+++ src/app.module.ts
@@ -2,7 +2,7 @@ import { Module } from "@nestjs/common";
 import { AppController } from "./app.controller";
 import { AppService } from "./app.service";
 import { AppConfigModule } from "./config/app/config.module";
-import { AuthModule } from "./auth/auth.module";
+import { AuthModule } from "./modules/auth/auth.module";
 
 @Module({
   imports: [AppConfigModule, AuthModule],
